app/inference/requirements.txt
jax
jaxlib
numpy
python-chess